refactor: divide `app.tsx` into smaller components
https://github.com/Attacktive/troubleshooter-editor-front-end/issues/216
