+++ .env
@@ -1,1 +1,2 @@
-BROWSER=none
+BROWSER = none
+REACT_APP_API_ROOT = http://localhost:8080
