+++ public/index.html
@@ -5,10 +5,7 @@
 	<link rel="icon" href="%PUBLIC_URL%/favicon.ico"/>
 	<meta name="viewport" content="width=device-width, initial-scale=1"/>
 	<meta name="theme-color" content="#000000"/>
-	<meta
-		name="description"
-		content="Web site created using create-react-app"
-	/>
+	<meta name="description" content="Troubleshooter Editor"/>
 	<link rel="apple-touch-icon" href="%PUBLIC_URL%/logo192.png"/>
 	<!--
 		manifest.json provides metadata used when your web app is installed on a
