--- src/constants.ts
@@ -1,1 +1,0 @@
-export const apiRoot = "http://localhost:8080";
